// src/app/main/authentication/model/LoginModel.ts
export class LoginModel {
    email: string;
    passwordHash: string;
    rememberMe: boolean;
    userId: number;
    userName: string;
    mobileNo: string;
    phoneNo: string;
    newPassword: string;
    oTP: string;
    applicationToken: string;
    URID: string| null;
    CorporateName: string | null;
}